settings: configure eslint and demo project
