--- .eslintrc.js
@@ -1,20 +1,0 @@
-module.exports = {
-  env: {
-    es6: true,
-    node: true
-  },
-  extends: ['plugin:@typescript-eslint/recommended', 'plugin:prettier/recommended', 'prettier/@typescript-eslint'],
-  globals: {
-    Atomics: 'readonly',
-    SharedArrayBuffer: 'readonly'
-  },
-  parser: '@typescript-eslint/parser',
-  parserOptions: {
-    ecmaVersion: 2020,
-    sourceType: 'module'
-  },
-  plugins: ['@typescript-eslint'],
-  rules: {
-    'prettier/prettier': ['error']
-  }
-};
