+++ gomshal/src/index.ts
