+++ gomshal/src/lib/gomshal.spec.ts
@@ -1,0 +1,5 @@
+import test from 'ava';
+
+test('openGoogleMapInPuppeteer', t => {
+  t.true('');
+});
