+++ package.json
@@ -22,20 +22,21 @@
   "scripts": {
     "help": "npm run && npm ls --depth 0",
     "test": "run-s build test:*",
-    "test:lint": "lint && prettier \"src/**/*.ts\" --list-different",
+    "test:lint": "lint",
     "test:unit": "nyc --silent ava",
-    "lint": "eslint '*/**/*.{js,ts,tsx}'",
+    "lint": "eslint .",
     "build": "run-s clean && run-p build:*",
     "build:main": "tsc -p tsconfig.json",
     "build:module": "tsc -p tsconfig.module.json",
+    "build:demo": "tsc -p tsconfig.demo.json",
     "fix": "run-s fix:*",
-    "fix:prettier": "prettier \"src/**/*.ts\" --write",
-    "fix:lint": "eslint '*/**/*.{js,ts,tsx}' --quiet --fix",
+    "fix:lint": "eslint . --quiet --fix",
     "watch": "run-s clean build:main && run-p \"build:main -- -w\" \"test:unit -- --watch\"",
     "cov": "run-s build test:unit cov:html && open-cli coverage/index.html",
     "cov:html": "nyc report --reporter=html",
     "cov:send": "nyc report --reporter=lcov && codecov",
     "cov:check": "nyc report && nyc check-coverage --lines 100 --functions 100 --branches 100",
+    "demo": "run-s build:demo && ",
     "doc": "run-s doc:html && open-cli build/docs/index.html",
     "doc:html": "typedoc src/ --exclude **/*.spec.ts --target ES6 --mode file --out build/docs",
     "doc:json": "typedoc src/ --exclude **/*.spec.ts --target ES6 --mode file --json build/docs/typedoc.json",
@@ -56,17 +57,15 @@
     "@istanbuljs/nyc-config-typescript": "^0.1.3",
     "@types/node": "^13.13.5",
     "@types/puppeteer": "^2.1.0",
-    "@typescript-eslint/eslint-plugin": "^2.32.0",
-    "@typescript-eslint/parser": "^2.32.0",
+    "@typescript-eslint/eslint-plugin": "^2.33.0",
+    "@typescript-eslint/parser": "^2.33.0",
     "ava": "^2.4.0",
     "codecov": "^3.6.5",
     "cz-conventional-changelog": "^2.1.0",
     "eslint": "^7.0.0",
-    "eslint-config-prettier": "^6.11.0",
     "eslint-config-standard": "^14.1.1",
     "eslint-plugin-import": "^2.20.2",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^3.1.3",
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.1",
     "gh-pages": "^2.2.0",
@@ -76,7 +75,6 @@
     "npm-run-all": "^4.1.5",
     "nyc": "^14.1.1",
     "open-cli": "^5.0.0",
-    "prettier": "^1.19.1",
     "standard-version": "^6.0.1",
     "trash-cli": "^3.0.0",
     "ts-node": "^8.10.1",
@@ -105,9 +103,6 @@
     "*.{js,ts,tsx}": [
       "eslint --fix"
     ]
-  },
-  "prettier": {
-    "singleQuote": true
   },
   "nyc": {
     "extends": "@istanbuljs/nyc-config-typescript",
