--- src/lib/gomshal.spec.ts
@@ -1,6 +1,0 @@
-// tslint:disable:no-expression-statement
-import test from 'ava';
-
-test('openGoogleMapInPuppeteer', t => {
-  t.true('');
-});
