+++ tsconfig.demo.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig",
+  "compilerOptions": {
+    "rootDir": "./demo/src",
+    "outDir": "build/demo",
+  },
+}
